requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==2.7.0
transformers==4.41.1
torch==2.3.0
faiss-cpu==1.8.0
numpy==1.26.4
langchain-text-splitters